Feat: Admin and User or Cashier Middleware

+++ routes/api.php
@@ -20,25 +20,25 @@ Route::middleware('auth:sanctum')->group(function () {
 });
 
 Route::prefix('/products')->middleware('auth:sanctum')->group(function () {
-    Route::post('/', [ProductController::class, 'createProduct']);
+    Route::post('/', [ProductController::class, 'createProduct'])->middleware('role:admin');
     Route::get('/', [ProductController::class, 'getProducts']);
     Route::get('/search', [ProductController::class, 'searchProduct']);
     Route::get('/low-stock', [ProductController::class, 'getLowStockProducts']);
     Route::get('/{id}', [ProductController::class, 'getProductById']);
-    Route::put('/{id}', [ProductController::class, 'updateProductById']);
-    Route::post('/{id}/adjust-stock', [ProductController::class, 'adjustStock']);
-    Route::delete('/{id}', [ProductController::class, 'deleteProductById']);
+    Route::put('/{id}', [ProductController::class, 'updateProductById'])->middleware('role:admin');
+    Route::post('/{id}/adjust-stock', [ProductController::class, 'adjustStock'])->middleware('role:admin');
+    Route::delete('/{id}', [ProductController::class, 'deleteProductById'])->middleware('role:admin');
 });
 
 Route::prefix('/stock-movements')->middleware('auth:sanctum')->group(function () {
-    Route::get('/', [StockMovementController::class, 'getStockMovements']);
+    Route::get('/', [StockMovementController::class, 'getStockMovements'])->middleware('role:admin');
 });
 
 Route::prefix('/transactions')->middleware('auth:sanctum')->group(function () {
     Route::get('/', [TransactionController::class, 'getTransactions']);
     Route::get('/{id}', [TransactionController::class, 'getTransactionById']);
     Route::post('/', [TransactionController::class, 'storeTransaction']);
-    Route::put('{id}/cancel', [TransactionController::class, 'cancelTransactionById']);
+    Route::put('{id}/cancel', [TransactionController::class, 'cancelTransactionById'])->middleware('role:admin');
 });
 
 Route::prefix('/dashboard')->middleware('auth:sanctum')->group(function () {
@@ -46,5 +46,5 @@ Route::prefix('/dashboard')->middleware('auth:sanctum')->group(function () {
 });
 
 Route::prefix('/reports')->middleware('auth:sanctum')->group(function () {
-    Route::get('/sales', [ReportController::class, 'getSalesReport']);
+    Route::get('/sales', [ReportController::class, 'getSalesReport'])->middleware('role:admin');
 });
